refactor: Extractor를 Parser에 통합
