+++ src/test/java/com/github/teamreflog/reflogserver/auth/infrastructure/TokenParserImplTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.github.teamreflog.reflogserver.auth.domain.Token;
-import com.github.teamreflog.reflogserver.auth.domain.TokenParser;
 import com.github.teamreflog.reflogserver.auth.exception.JwtInvalidException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
@@ -22,7 +21,7 @@ class TokenParserImplTest {
     static final SecretKey SECRET_KEY =
             Keys.hmacShaKeyFor("asdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdfasdf".getBytes());
 
-    TokenParser parser;
+    TokenParserImpl parser;
 
     @BeforeEach
     void setUp() {
@@ -38,7 +37,7 @@ class TokenParserImplTest {
         void returnJwt() {
             /* given */
             final String token =
-                    "eyJhbGciOiJIUzM4NCJ9.eyJzdWIiOiIxIiwicm9sZSI6Ik1FTUJFUiIsImV4cCI6OTk5OTk5OTk5OX0.gBhNpEccOhdlwNRg3jcVGTw_fXLY2nKkObcp6P_tUqvbKQvGiLcfGpQzmP9rEPbo";
+                    "Bearer eyJhbGciOiJIUzM4NCJ9.eyJzdWIiOiIxIiwicm9sZSI6Ik1FTUJFUiIsImV4cCI6OTk5OTk5OTk5OX0.gBhNpEccOhdlwNRg3jcVGTw_fXLY2nKkObcp6P_tUqvbKQvGiLcfGpQzmP9rEPbo";
 
             /* when */
             final Token jwt = parser.parse(token);
@@ -65,7 +64,7 @@ class TokenParserImplTest {
         void throwExceptionWithExpiredJwt() {
             /* given */
             final String expired =
-                    "eyJhbGciOiJIUzM4NCJ9.eyJzdWIiOiIxIiwicm9sZSI6Ik1FTUJFUiIsImV4cCI6MTcwNTAwMDAwMH0.WnDG4XnyKX9nwl1W1MXggvaj43231oqgxADB4PCzo32RoKReDpMhs3Xpt82jcNbe";
+                    "Bearer eyJhbGciOiJIUzM4NCJ9.eyJzdWIiOiIxIiwicm9sZSI6Ik1FTUJFUiIsImV4cCI6MTcwNTAwMDAwMH0.WnDG4XnyKX9nwl1W1MXggvaj43231oqgxADB4PCzo32RoKReDpMhs3Xpt82jcNbe";
 
             /* when & then */
             assertThatCode(() -> parser.parse(expired))
@@ -93,12 +92,44 @@ class TokenParserImplTest {
         void throwExceptionWithNotExistClaim() {
             /* given */
             final String noSubject =
-                    "eyJhbGciOiJIUzM4NCJ9.eyJyb2xlIjoiTUVNQkVSIiwiZXhwIjo5OTk5OTk5OTk5fQ.fpZCMUnDn5HdBsYsYQC4HYwHnrQManIEfKbYN12QNBqFlxWqCL9hkO7pK6u6oeFA";
+                    "Bearer eyJhbGciOiJIUzM4NCJ9.eyJyb2xlIjoiTUVNQkVSIiwiZXhwIjo5OTk5OTk5OTk5fQ.fpZCMUnDn5HdBsYsYQC4HYwHnrQManIEfKbYN12QNBqFlxWqCL9hkO7pK6u6oeFA";
 
             /* when & then */
             assertThatCode(() -> parser.parse(noSubject))
                     .isInstanceOf(JwtInvalidException.class)
                     .hasMessage("유효하지 않은 토큰입니다.");
+        }
+
+        @Test
+        @DisplayName("토큰이 Bearer로 시작하면 토큰을 추출한다.")
+        void extractBearerToken() {
+            /* given */
+            final String bearerToken = "Bearer ekfrEjrqhRdlaktdlTrpTek";
+
+            /* when */
+            final String token = parser.extract(bearerToken);
+
+            /* then */
+            assertThat(token).isEqualTo("ekfrEjrqhRdlaktdlTrpTek");
+        }
+
+        @Test
+        @DisplayName("토큰 타입이 Bearer가 아니면 예외가 발생한다.")
+        void throwExceptionWithWrongJwtType() {
+            /* given */
+            final String noType = "ekfrEjrqhRdlaktdlTrpTek";
+            final String wrongType = "Wrong ekfrEjrqhRdlaktdlTrpTek";
+
+            /* when */
+            assertAll(
+                    () ->
+                            assertThatCode(() -> parser.extract(noType))
+                                    .isInstanceOf(JwtInvalidException.class)
+                                    .hasMessage("유효하지 않은 토큰입니다."),
+                    () ->
+                            assertThatCode(() -> parser.extract(wrongType))
+                                    .isInstanceOf(JwtInvalidException.class)
+                                    .hasMessage("유효하지 않은 토큰입니다."));
         }
     }
 }
